refactor: renaming refactor to /ContentEditableInsert
