+++ src/handlers/ContentEditableInsert.js
